LayoutTests/imported/w3c/web-platform-tests/css/css-align/blocks/align-content-block-007-expected.txt

FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
FLT
FLTABSREL
OVERFLOW
Show Text


PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
FAIL .test 9 assert_equals:
<div class="test" style="align-content: unsafe center" title="unsafe center">
    <div class="float" data-offset-y="25">FLT</div>
    <div class="in-flow" data-offset-y="35"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="35">FLT</div>
      <span class="label">UNSAFE CENTER</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 25 but got 15
FAIL .test 10 assert_equals:
<div class="test" style="align-content: unsafe end" title="unsafe end">
    <div class="float" data-offset-y="15">FLT</div>
    <div class="in-flow" data-offset-y="25"></div>
    <div class="in-flow">
      <div class="float" data-offset-y="25">FLT</div>
      <span class="label">UNSAFE END</span>
      <span class="abspos">ABS</span>
      <span class="relpos">REL</span>
      <div class="overflow">OVERFLOW</div>
    </div>
  </div>
offsetTop expected 15 but got -5
PASS .test 11
PASS .test 12
PASS .test 13
PASS .test 14
PASS .test 15
PASS .test 16
PASS .test 17
